Additional refactors and modularizations. Review dex-pokemon, dex-movesets.
